# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de autenticação",
        "title": "Auth-Go API",
        "contact": {},
        "version": "1.0"
    },
    "host": "auth.127.0.0.1.lvh.me",
    "basePath": "/",
    "paths": {
        "/admin": {
            "get": {
                "description": "Endpoint restrito a admin",
                "tags": [
                    "admin"
                ],
                "summary": "Painel admin",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Autentica o usuário e retorna o token JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Realizar Login",
                "parameters": [
                    {
                        "description": "Credenciais de Login",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token JWT",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "server.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}